Add lobby system with join codes
