+++ public/index.html
@@ -10,7 +10,15 @@
     <h1>Larry &amp; Barney's Monopoly</h1>
     <div id="join">
         <input id="name" placeholder="Your name" />
-        <button id="joinBtn">Join Game</button>
+        <div>
+            <input id="lobbyName" placeholder="Lobby name" />
+            <input id="createCode" placeholder="Join code" />
+            <button id="createBtn">Create Lobby</button>
+        </div>
+        <div>
+            <input id="joinCode" placeholder="Join code" />
+            <button id="joinBtn">Join Lobby</button>
+        </div>
     </div>
     <div id="game" style="display:none;">
         <div id="layout">
